+++ frontend/src/components/helper/auth-redirect.tsx
@@ -21,23 +21,3 @@ export function withAuthRedirect(WrappedComponent: ComponentType<any>) {
     return <WrappedComponent {...props} />;
   };
 }
-
-// New HOC to prevent logged-in users from accessing the login page
-export function withLoggedOutRedirect(WrappedComponent: ComponentType<any>) {
-  return function LoggedOutRedirectWrapper(props: ReactElement) {
-    const navigate = useNavigate();
-    const query = useAuthUser();
-
-    if (query.isLoading) {
-      return <div className="text-3xl font-bold">Loading... with tailwind</div>;
-    }
-
-    const user = query.data?.data?.user;
-    if (user) {
-      navigate({ to: "/" });
-      return null;
-    }
-
-    return <WrappedComponent {...props} />;
-  };
-}
